requirements.txt
python-dotenv==1.1.0
pytelegrambotapi==4.27.0
aiohttp==3.11.18